feat: Add install command
Added a new install command to the CLI application. This command allows users to install packages. The command calls the `Install` function in the `commands` package, which in turn calls the `CreateCacheDotFolder` function in the `module` package to create a cache dot folder. This commit adds a new file `install.go` in the `commands` package and modifies the `main.go` and `cache.go` files.

Changes:
- Added `install.go` file in the `commands` package with the `Install` function.
- Modified `main.go` to include the `install` command and call the `commands.Install` function.
- Renamed `cache.go` to `module/cache.go`.
- Modified the `CreateCacheDotFolder` function in `module/cache.go` to be exported (`CreateCacheDotFolder` instead of `createCacheDotFolder`).